Adding `uninstall.php` stub file; together with `uninstall.inc.php` uninstall handler. See: websharks/comment-mail#12
