+++ comment-mail/plugin.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * Plugin
+ *
+ * @package plugin
+ * @since 14xxxx First documented version.
+ * @copyright WebSharks, Inc. <http://www.websharks-inc.com>
+ * @license GNU General Public License, version 3
+ */
 /*
 Version: 14xxxx
 Text Domain: comment-mail
